fiddled with the long vs. int (short, byte) problems, all test cases working

+++ hu.bme.mit.incqueryd.core/src/main/java/hu/bme/mit/incqueryd/rete/comparison/Compare.java
@@ -6,7 +6,7 @@ package hu.bme.mit.incqueryd.rete.comparison;
  *
  */
 public class Compare {
-    public static boolean compare(final int a, final int b, final ComparisonOperator comparisonOperator) {
+    public static boolean compare(final long a, final long b, final ComparisonOperator comparisonOperator) {
         switch (comparisonOperator) {
         case EQUALS:
             return a == b;
